fix: use merge instead of object spread (#1231)
### Issue

It is very easy to merge two JavaScript objects incorrectly using the spread operator. Typically when we want merge some defaults and overrides, we use `const result = { ...defaults, ...overrides }`. While it works, it may not produce the expected value if any property of `overrides` is `undefined`.

For all _missing_ properties in the `overrides` object, the functionality is correct. But if there is an object property and the value of that property is `undefined`, the merged object doesn't fallback to the default value. 

E.g.
```ts
interface SomeOptions {
    a: string
}
const defaultValues = {
    a: 'foobar'
}
const overrideValues: Partial<SomeOptions> = {
    a: undefined,
}
const result = {
    ...defaultValues,
    ...overrideValues
}
console.log(result) // { a: undefined }
```

Any property which is annotated with question mark (directly, or via `Partial`) can have explicit value of `undefined`.

Note also that e.g. in VS Code and IntelliJ IDEA the inferred type of `result` is shown as:
```ts
{
    a: string;
}
```

(There is a TypeScript configuration option to control this: https://www.typescriptlang.org/tsconfig#exactOptionalPropertyTypes)

### Fix

We need to use something else than the spread operator.

In this PR we added a new `merge` utility function to `@streamr/utils`. It does a simple non-recursive merge and doesn't mutate any of the argument objects. But it skips all `undefined` properties.

Alternatively we could use `merge` from `lodash` if recursive merge is needed (and if mutating the target argument is ok).

With both methods we get valid merged object:
```ts
console.log(merge(defaultValues, overrideValues))  // { a: 'foobar' }
```