feat(client): [NET-968] Missing partition count in metadata (#1240)
When we parse stream data, we fallback to partition count 1 if there is no information about the partition count in the contract metadata:
- if the metadata JSON doesn't have `partitions` field
- if the metadata is not a JSON

In all other cases the functionality is the same as previously: i.e. invalid partition count throws an error. That error is now `StreamrClientError` instead of generic `Error` so that we can provide an error code to end users.