Docs (#1230)
## Summary

Merged the old documentation repository (Docusaurus docs) into the
monorepo

## Changes

Removed the old ./git folder inside of the old documentation folder. Git
flow now works with this repository.

Added the docs to root/docs after consultation with Eric. 

Renamed old "docs" folder in root repository to "network-docs"

## Limitations and future improvements

We need to add automation for deployment to prod.

## Note for reviewer:
You can ignore all the markdown and yml files
Please focus mostly on the merge of the repository. Is the folder
structure right, in your opinion?

---------

Co-authored-by: Eric Andrews <harbu88@hotmail.com>
Co-authored-by: Jarno <jarno.marttila@sujuwa.fi>
Co-authored-by: Lasse <git@crunchybar.com>