fix(network): do not send duplicate messages when buffers full (#1290)
## Summary

Fixes an issue in the `network` package in which duplicate messages
would be sent out when the send buffer of a WebRTC connection (on
Node.js) had items in it.

## Detailed explanation

The root of this bug is in how our code interpreted the returned boolean
value of method `dataChannel#send` of `node-datachannel`. Our code
assumed that this boolean value indicates the success of sending a
message. However, it actually indicates whether the message could be
sent immediately or whether it had to be buffered by the library (see
https://github.com/murat-dogan/node-datachannel/issues/159). So when the
return value was `false`, our code would re-send the same message later
(up to 10 times) even though the message was already accepted by the
library to be sent when its internal buffer cleared a bit.

Another side-effect of this bug: we may have warn logged messages as not
having been sent even though they probably were (including duplicates).

Fixed this issue by ignoring the return value. We already implement low
/ high buffer handling on our end (with the assistance of
`dataChannel#setBufferedAmountLowThreshold`.