test(cli-tools): Add end-to-end tests (#1301)
## Summary

Add end-to-end test for `cli-tools` package. Most functionality is now under tests.

## Implementation

We build the `cli-tools` package before running the tests. It allows us to run the scripts with `node` instead of e.g. `ts-node`. Alternatively we could install `ts-node` as a dev dependency and run scripts like this: `../../node_modules/.bin/ts-node bin/streamr-wallet-whoami.ts`. But with that approach the startup of the interpreter is significantly slower (~30s).

## Limitations and future improvements

- Add tests for the `governance vote` command
- Improve test coverage so that all command line arguments have some test case
- Move `collect` to `@streamr/utils` package and also refactor client to use this shared function

## Checklist before requesting a review

- [ ] Is this a breaking change? If it is, be clear in summary.
- [ ] Read through code myself one more time.
- [ ] Make sure any and all `TODO` comments left behind are meant to be left in.
- [ ] Has reasonable passing test coverage?
- [ ] Updated changelog if applicable.
- [ ] Updated documentation if applicable.