fix(client): Re-assign storage node (#1302)
## Summary

If there was an existing storage node assignment, and `stream.addToStorageNode()` was called again for that same storage node, the operation timeouted.

## Changes

We explicitly check whether the stream is already stored. The assignment event listener logic requires that there must not be an existing assignment (as the storage node doesn't send an assignment event in that case).