NET-974: external ip (#1357)
## Summary

Make it possible to configure external IPs for the WebRTC connections.
This is useful in cases where the peer has a public IP address but the
host does not know of it. This happens in cases where the node is behind
a Full Cone NAT. For example, in AWS it is common that the host's
network interface does not know it's public IP.

The external IP address is injected into WebRTC IceCandidate messages
when the type is known to be `host`, making connecting to such peers
possible without the use of STUN.

## Changes

- Added configurability for NodeJS WebRTC connections for an external IP
address
- Added configurability all the way to the client

## Limitations and future improvements

- The solution should not work in cases where the the Node is behind a
Full Cone NAT that randomizes the ports that are used. It works only if
the port local port mappings are preserved on the public side.

## Checklist before requesting a review

- [ ] Is this a breaking change? If it is, be clear in summary.
- [ ] Read through code myself one more time.
- [ ] Make sure any and all `TODO` comments left behind are meant to be
left in.
- [ ] Has reasonable passing test coverage?
- [ ] Updated changelog if applicable.
- [ ] Updated documentation if applicable.