test(client): Simplify gap fill tests (#1367)
Replace integration test with new unit tests and a simple integration test. The previous test used monkey patching of subscription pipeline, and was therefore it relied too much to the current implementation details.

### Open questions

In the previous integration test there was a test case for doing gap fill for resends subscriptions. Do we support that feature? Or should we remove the feature? If we support it, we may want to add similar test case to the new integration test.