fix(broker): Default value for resend range queries (#1461)
Fixed default value handling of `toSequenceNumber` in the `storage` plugin resend range handler. 

The previously used `||` operator handled 0 value incorrectly. As value `0` is one of the _falsy_ values in JavaScript, if we provided explicit value of `0`, it used the default of `MAX_SEQUENCE_NUMBER_VALUE` instead.

Changed also two other use cases of `|| `operator in `dataQueryEndpoints`. That change doesn't change the functionality as the default is `0` in those cases.

### Related issue

Also note a related client fix of another bug https://github.com/streamr-dev/network/pull/1462. When we deploy this fix to storage node, the resend range functionality in old clients break because the clients have a bug.