fix(client): Default value for resend range queries (#1462)
The default value of `toSequenceNumber` for resend range queries was incorrectly set to `0`.

Removed the default value handling of this query parameter, because the parameter is optional. The `storage` plugin applies the default value if we don't specify any query parameter value.

Removed also the default value of two other optional query parameters. That change doesn't change the functionality as the default is `0` in those cases.

### Related issue

Note that the default value which `storage` plugin uses is currently incorrect. PR https://github.com/streamr-dev/network/pull/1461 fixes that.