fix(client): Pipeline error handling (#1479)
Added error handling  block to `messagePipeline.ts`. There was no `try-catch` in the outermost message iteration loop. If some component wrote an explicit error to a pipeline (with `pipeline.endWrite(err)`), the pipeline didn't handle that.

E.g. when we do a resend, `HttpUtils#fetchHttpStream` writes error to the pipeline if a `fetch` call throws:
- `fetchHttpStream` throws an error to the returned `AsyncIterable`
- `pull` function catches it in `PushBuffer.ts#265` 

Therefore it was possible that a network error during threw an unhandled promise rejection error which crashed the application process.