NET-939:  Proxy pattern, can open and close connections with data flow. (#1485)
## Summary

The Proxy Pattern's MVP implementation. The connections work and data is
flowing through as expected. There is yet no reconnection logic
implememented.

The ProxyStreamConnectionClient is a parallel implementation for a
Layer2 node. Thus the PR includes required changes and new interfaces to
make this possible.

Next up the reconnection logic, Layer0 join avoidance and client
integration will be implemented to complete the feature.