refactor(client): Extract resend pipeline (#1473)
Extract resend pipeline from ResendSubscription.

In this PR there is no longer ResendSubscription, but instead resend subscription pipeline is initalized by calling initResendSubscription. With this approach we have only one Subscription class, which is used for both realtime and resend subscriptions. The main difference between realtime and resend subscriptions is the pipeline: in resend subscription there are some piped steps in the Subscription's internal pipeline, whereas for realtime subscription just uses an empty pipeline.

This PR a milestone so that we can implement raw support for resend subscriptions.

We should maybe later have even more separation between subscription and the producing pipeline. E.g. the resend subscription could maybe work similarly to realtime subscription in that sense that there is a producer which pushes data to the internal pipeline of the Subscription (see SubscriberSession line 67).