refactor(client): Message pipeline is `PushPipeline` (#1489)
Refactored message pipeline so that it is seen as `PushPipeline` instead of `MessageStream`. The `MessageStream` is a wrapper class for public API which allows iterating items as `Message` objects. 

When we process messages internally, we iterate `StreamMessage` objects, and therefore don't need the wrapper.

This way we can also remove some internal methods from `MessageStream`, which just delegated calls to `PushPipeline`.