refactor(client): The Graph timeout config (#1498)
Renamed/moved timeout config options for The Graph. 

Changes:
- `_timeouts.httpFetchTimeout` -> `_timeouts.theGraph.fetchTimeout`
- `_timeouts.theGraph.timeout` -> `_timeouts.theGraph.indexTimeout`
- `_timeouts.theGraph.retryInterval` -> `_timeouts.theGraph.indexPollInterval`

If makes sense to move `_timeouts.httpFetchTimeout` option into `theGraph` block as is not used by other components.

This is not a breaking change because `_timeouts` config block is not part of public API.

Removed also `HttpFetcher` class which is no longer needed. Also the default value of `httpFetchTimeout` was removed config `CONFIG_TEST` as it is same as in production config.