test(client): Fake storage node is a server (#1503)
Before this PR `FakeStorageNode` used `FakeHttpUtils` to serve stored messages. It didn't really do HTTP requests, but instead just called methods of  `FakeStorageNode`.  It was convenient implementation as the client already had a `HttpUtils` service in DI.

In this PR we a real `express` server instead of mocking HTTP requests. Each storage node launches a separate server when the node is started and destroys it when node is stopped. This way the tests coverage is better as more functionality of storage nodes are actually tested. 

The best coverage would be achieved if we use production environment or docker-dev environment. But fake environment testing is very useful because:
- we can simulate error scenarios
- we can be define more precisely when messages are stored
- integration tests are significantly faster than end-to-end tests

## Future improvements

- This change allows us to simplify client dependency injections as we can now remove `HttpUtils` service.