refactor(client): Remove `HttpUtils` (#1512)
Remove `HttpUtils` dependency which is no longer needed. 

It was used only by `Resends`: moved most of the logic to that class, and some logic to `utils.ts`.

Removed some low level logging from `fetchResponse`. We already have some `debug`-level logging in `Resends#fetchStream` which logs enough information for successful fetch operations. If there is an error, the relevant information is stored into the `Error` instance. Timing information is not needed as `fetch` provides the output as a stream.

## Future improvements

- Maybe some of the moved logic is not needed in `Resends`. E.g. error handling could be simpler (no need for separate class, and maybe we could use `StreamrClientError` class there)