NET-846: Fix flakyness in finding data under churn (#1515)
## Summary

Data can now be found more reliably from the DHT store if churn is
present. The recursive finds with data-mode on will traverse to the
closest peers in all cases. Additionally data that is stored on nodes
that are too far to the closest peers to it is marked as stale.