NET-927: Use the Trackerless network in client (#1037)
## Summary

Please provide a summary of the changes and a motivation if applicable.

## Changes

Provide a bullet list of individual changes. Leave this section out if
change
set is small and obvious from summary.

## Limitations and future improvements

Provide a bullet list or description of known omissions or limitations
of the
solution and/or any ideas for future improvement. Leave section ouf if
not applicable.

## Checklist before requesting a review

- [ ] Is this a breaking change? If it is, be clear in summary.
- [ ] Read through code myself one more time.
- [ ] Make sure any and all `TODO` comments left behind are meant to be
left in.
- [ ] Has reasonable passing test coverage?
- [ ] Updated changelog if applicable.
- [ ] Updated documentation if applicable.

---------

Co-authored-by: Eric Andrews <eric.andrews@streamr.com>