NET-846: Trackerless Network (#697)
<!--- Provide a general summary of the changes -->
## Trackerless Network

Base branch for the Trackerless Network using the Streamr DHT for peer
discovery, control traffic routing and connection management. The PR
includes supporting changes to the proto-rpc and DHT packages as well as
the package for the Trackerless Network.

A general view of the Overlay Networking stack for guidance:

![overlay-network-stack
drawio](https://user-images.githubusercontent.com/17498529/216920962-941ad08e-fbdc-4977-a1e7-8649aad82abc.png)

## Overview of changes to existing packages

### Proto-RPC changes

Start using Protobuf type `Any` for message values that can be of
multiple types, instead of using binary packaging with a separate type
field. (This pattern is more generally applied throughout the other
packages as well.)

The RpcCommunicator now has clearer differentiation between client and
server side errors

### Dht Changes

The DhtNode and the ConnectionManager now implement many reliability,
fault tolerance and performance improvements that make the system
stabler for real world applications. (ie. rejoin DHT, reattempt
ConnectivityChecks, routing performance upgrades, graceful leave
notifications)

**DhtNode:**

The SortedContactList class now has a alternate structure called
RandomContactList to which peers are inserted into on a random basis.
Both classes now extend the EventEmitter3 library and which makes
PeerDiscovery possible on the higher layers of the Overlay Network
stack.

Many operations performed by the DhtNode are now separated into their
own `Session`-files. These sessions are used atomically to perform tasks
such as routing a message (RoutingSession), finding peers in the
(DiscoverySession, RecursiveFind)

When The nodes join the network in two parallel ways. As before, firstly
the nodes join by traversing the Kademlia DHT from the entrypoint
towards their own KademliaId. Secondly, the nodes generate a random ID
and traverse the Kademlia DHT towards it. By doing so a node will find
peers that are more distant to it which enables faster generation of
Kademlia's redundant paths.

DhtRouting duplicate detection is now done using a Set based detection
algorithm instead of Bloom filters to avoid false positives.

Dht Routing is now performed recursively based on the connections on the
lower layer instead of traversing the Kademlia DHT.

**ConnectionManager:**

The simulator now more accurately emulates real world conditions. This
includes latency conditions and implementation details when compared to
the WebSocket and WebRTC Connectors.

When connections are opened simultaneously from both ends one of the
connections is killed with logic that is consistent on both ends. This
reduces packet loss and network level duplicates.

`disconnected` and `connected` events are now emitted for the higher
layers of the networking stack.

Connections can now be hard locked by services on the higher layers.
Hard locked connections are locked on both ends of the connection. This
ensures that the Connection Manager does not clean up connections that
are used by ie. the Layer2 RandomGraphNodes.

Connections soft locked connections. These are connections which are
only locked on one end of the connections. Connections are soft locked
when they are added to the K-Buckets on the DhtNodes.

The Connection Manager has a soft limit for the number of connections.
If the number of connections is above the limit of 80 the manager will
clean up unused and unlocked connections until the number of connections
is at the limit.

### Monorepo level dependency changes

The individual packages no longer use the `@protobuf-ts/plugin`. Instead
they only use what ever is needed during runtime. Moreover, the Protobuf
generate code is no longer generated during `npm run build`. If there
are changes to the generated code, the files should be regenerated with
`npm run generate-protoc-code` and pushed to the remote repository.
These changes were made to fix build errors in the Linux environments.


### Trackerless Network

The Trackerless Network package is a reimplementation of the old
(Corea-Brubeck) network package. The package is reimplemented to use the
network and transport stacks of the proto-rpc and DHT packages. The main
change to the network is that the `d-regular random graph` stream
topologies are now generated using a decentralized algorithm based on
peer discovery from the DHT.


**NetworkNode:**

The Trackerless Network implements a similar interface to the old
Network's `NetworkNode` class which is used as an interface for the
StreamrClient. For now, there is built in integration to translate the
old Network's `streamr-client-protocol`
messages to the new protobuf based messages. In the future, the client
should move to use protobuf.

**NetworkStack:**

The `NetworkStack` class is used to generate and start compositions of
Streamr Network Nodes.

**StreamrNode:**

The `StreamrNode` class is used to join, leave, broadcast and listen to
topics (Streams) in the network. The class generates compositions of
Layer2 - Layer1 for each Stream it knows.

**Layer0:**

The Layer 0 in the network is a DhtNode which uses a ConnectionManager
as its transport layer. The Layer0 is used to join the network, to
discover, find peers closest to an id and to route messages (ie. webrtc
signaling messages) to peers in the network.

**Per Stream Peer Discovery (Layer1):**

The Layer1 nodes in the network are virtual DhtNodes. This means that
they do not form their own connections and instead they use the Layer0
DhtNodes as their transport layer. In practice, each `StreamPartition`
has its own network of Layer1 DhtNodes that is used to join the Stream
and to find peers in the Stream.

**RandomGraphNode (Layer2):**

The `RandomGraphNode` class is the implementation for Layer2 in the
network. Each node attempts to open connections to a target number of
peers which it discovers using a Layer 1 DhtNode. In practice, each
`StreamPartition` has its own network of Layer2 nodes that is used to
propagate content traffic.

The Layer2 nodes use the ConnectionManager to open and lock direct P2P
connections to its neighbors.

## Checklist

<!--- Feel free to remove any that are not relevant -->

- [x] Has passing tests that demonstrate this change works
- [ ] Updated Changelog
- [ ] Updated Documentation

---------

Co-authored-by: Petri Savolainen <petri.savolainen@spaceify.net>
Co-authored-by: Eric Andrews <eric.andrews@streamr.com>