refactor(client): Return type of `fetchHttpStream` (#1552)
The `fetchHttpStream` utility now returns lines instead of `StreamMessage` objects. The core functionality of that function is to fetch data from a streaming HTTP endpoint: it is not relevant what kind of data it is handling.

The transformation from lines to messages is done in the caller method (`Resends#fetchStream`). The iteration loop of that function has also been simplified (by removing the `counting` utility function). The pulling also launches the background task explicitly by calling `setImmediate`.

Added also a test case for checking that large resends are handled correctly.