refactor(client): Inline `PushPipeline#pull` method (#1553)
Inlined `PushPipeline#pull` method which was just a delegated to the `pull` function in `PushBuffer`. The utility function works both with `PushPipeline` and `PushBuffer` it would make sense to have the delegator method in both classes, or in neither class. Maybe the best approach is to keep the pipeline and buffer classes as simple as possible.

Removed also `PushBuffer#pull` and `PushBuffer#filter` methods which were not used.