fix(client): Error handling of `fetchHttpStream` (#1554)
If the HTTP `fetch` stream fails, it now propagates the error to the piped stream (the stream which transforms chunks of texts to lines). The underlying stream throws an error e.g. if we call `AbortController#abort()`. 

Before this PR the error in the underlying stream would just stop the output of that stream, but the transform stream didn't stop as it was not notified by the error. The `AsyncGenerator` never completed in that case.

Also replaced `parseError` functionality: the `fetchHttpStream` now throws an explicit `FetchHttpStreamResponseError` if the response is not ok (i.e. HTTP 200) instead of calling `parseError` to form the error.