NET-975: stored data can be marked as  deleted in the dht (#1572)
## Summary

Stored data can now be marked as deleted in by its original storer in
the DHT. The RecursiveFindSession will now check if the latest stored
data has been marked as deleted before passing it on to the requestor.

The stored data's storer's timestamp has been included to the
dataEntry-objects and is used when handling the data. This ensures that
the latest data sent by the storer will be kept in storage.