refactor(client): Remove `StreamRegistryCached' (#1588)
Moved all logic from `StreamRegistryCache` to `StreamRegistry`. This way the caching is encapsulated into the `StreamRegistry` class. The users of the stream registry can just call normal methods of `StreamRegistry` and the class handles the caching automatically.

This PR is a milestone to add caching to other registries, e.g. `StreamStorageNodeRegistry`. It is not convenient to have a separate caching wrapper for each of our service class which uses caching.

### Future improvements

- The `StreamRegistry#clearStreamCache` clears all other caches, but not the cache for `hasPublicSubscribePermission`. Maybe it should clear that, too.
- Currently the methods of `StreamrClient` don't use cache when `getStream` / `isStreamPublisher` / `isStreamSubscriber` is called. Maybe would be make sense to use cache? (if we do that, we don't maybe need `useCache` parameters)