feat(client): [NET-865]: gap filling strategy option (#1628)
## Summary

Add the ability to control the gap fill strategy with `gapFillStrategy` config option.

By default this is `light` to be in line with current behavior. However, if more pedantic gap filling is wanted by the user, it is now possible to enable it by setting the option to `full`.

## Open questions

Would it make sense to unify `gapFillStrategy` and the the `gapFill` boolean flag? E.g. `gapFillStrategy='passive'` would correspond to current `gapFill=false`. Maybe `orderMessages` flag shouldn't be a separate flag?

## Benchmark

Typically the new implementation does fewer resend requests compared to the previous implementation (the requests cover larger message ranges). That is most likely the root cause why the `OrderMessages2.test.ts` unit test is slower than the similar `OrderingUtil2.test.ts` in In PR https://github.com/streamr-dev/network/pull/732. The new tests takes about 23 seconds to run, whereas the old tests took takes about 19 seconds to run (i.e. the the new one takes is 20% more time). 

As `OrderMessages2.test.ts` test is quite slow, the test  is currently disabled (by Jest `skip`). It would be good to run it manually when we make changes to ordering functionality.