feat(network): NET-1001 Simplified Inspect (#1728)
## Summary

New feature for nodes to `Inspect` other nodes in stream partitions. The
inspecting nodes will monitor that the inspected node is propagating
data by opening a connection to the inspected node. The inspection will
succeed once any message is received from the rest of the stream and the
inspected node. Inspection will also succeed if the inspection timeouts
and there were no messages received during the inspection.

## Future improvements

- Client integration
- Run FindPeer and Inspect in the same operation?
- Limit the size of the `inspectionMessages` map in InspectSession.ts
- Optimize calling of `Inspect.markMessage()`, currently it is called
whenever the StreamNodeServer receives a new stream message.