refactor: extract `isLeader` helper function (#1754)
## Summary

Extract `createIsLeaderFn` function for creating a function to check for
leader. The same code can be used in all the places where we need to
check for leader.

Also switch `isHeartbeatStale() && isLeader()` to `isLeader() &&
isHeartbeatStale()` for efficiency. Only the leader needs to actually go
to the graph and query what the last heartbeat timestamp was.