feat(network): Websocket port range (#1801)
## Summary

The network layer WebSocket servers are now started in a given port
range. By default this range is between 32200 and 32250 for production
values. For development the port range is set between the ports 32400
and 32800.

When starting a node the range of ports is iterated through to find an
open port. Once a open port is found the node will initiate connectivity
checks as before. If an open port is not found the node will continue to
operate without a WS server and will be dependant on connecting on WS
client and WebRTC connections.

Starting the servers in avoided in browser environments to avoid
crashing behaviour.

## Additional Changes

- Fixed a leaking test case in Broker where the `stop()` was never
called.
- Fixed leaking test where a client was never destroyed.
- Added safe guards to WebSocket connection cases where leaks might
occure (The observed WS related leaks are likely caused by stop not
being called in all tests in the broker and client projects.)

## Limitations and future improvements

- A new network control layer option to completely avoid starting a
WebSocket server could be added.

- (The client and broker tests the `afterEach` blocks often call
stop/destroy on the client or broker with the `?` operator. In cases
where timeouts occure during start this will cause stop not to be
called. This causes logging to leak on top of other tests and in the
worst cases other test cases will fail because of it. It would be better
to use the `!` operator to cause these test cases to fail and tear down
their environments in the process instead of allowing failed test cases
to leak on top of others.)