test(client): [NET-1022]: Fix flaky `multiple-clients` test (#1817)
The test was flaky, because we have concurrency bug (or missing feature) in encryption handling: if two clients open the local DB at the same time, it is possible that one of the client fails to open the DB as it has not "migrated" yet (i.e. the tables haven't been created). There is a separate ticket about this bug NET-1057, and in this PR we added a new test which focuses to this bug. The new test is not flaky as it always fails (and therefore it is currently skipped). 

The actual flaky test (`multiple-clients.test.ts`) has been resolved by using separate private keys for each subscriber. The feature under test here is more abstract, and the private keys of the subscribers are not relevant to this test. The main idea in the flaky test is to check whether there can be multiple publishers on a single stream. It could be implement with just with one subscriber, or like we do now, with two subscribers having separate private keys.